# lib/apispec.json
{
  "swagger": "2.0",
  "info": {
    "title": "MetaDisk API"
  },
  "host": "api.metadisk.org",
  "basePath": "/",
  "schemes": ["https"],
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "securityDefinitions": {
    "basic": {
      "type": "basic",
      "description": "HTTP Basic Authentication with email:sha256password",
      "name": "Authorization",
      "in": "header"
    },
    "ecdsa": {
      "type": "apiKey",
      "description": "ECDSA pubkey header and x-signature of the request body including nonce",
      "name": "x-pubkey",
      "in": "header"
    }
  },
...
  "paths": {
    "/users": {
      "post": {
        "tags": ["users"],
        "summary": "Registers a new user account with MetaDisk",
        "parameters": [
          {
            "name": "user",
            "in": "body",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "email": {
                  "description": "Email address for account activation and login",
                  "required": true,
                  "type": "string"
                },
                "password": {
                  "description": "SHA-256 hash of your plaintext password",
                  "required": true,
                  "type": "string"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "User successfully registered and activation email sent",
            "schema": {
              "type": "object",
              "properties": {
                "email": {
                  "type": "string",
                  "example": "gordon@storj.io"
                },
                "created": {
                  "type": "string",
                  "example": "2016-03-04T17:01:02.629Z"
                },
                "activated": {
                  "type": "boolean",
                  "example": false
                }
              }
            }
          }
        }
      }
    },
    "/activations/{token}": {
      "get": {
        "tags": ["users"],
        "summary": "Activates a registered user and optionally redirects",
        "parameters": [
          {
            "name": "token",
            "in": "path",
            "type": "string",
            "description": "Activation token sent to user's email address",
            "required": true
          },
          {
            "name": "redirect",
            "in": "query",
            "type": "string",
            "description": "Optional redirect URL for successful activation"
          }
        ],
        "responses": {
          "200": {
            "description": "User successfully activated",
            "schema": {
              "type": "object",
              "properties": {
                "email": {
                  "type": "string",
                  "example": "gordon@storj.io"
                },
                "created": {
                  "type": "string",
                  "example": "2016-03-04T17:01:02.629Z"
                },
                "activated": {
                  "type": "boolean",
                  "example": true
                }
              }
            }
          }
        }
      }
    },
...
  }
}